# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "creationGoal-api",
    "description": "CreationGoal Process App Personal Project",
    "version": "1.0.0"
  },
  "host": "p2-web-services-w4.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/api-docs": {
      "get": {
        "summary": "GETS the SWAGGER API DOCUMENTATION page ____________*** This api-docs page is in sync with the ---(OAUTH AUTHORIZATION)--- from the LOGIN entry page***",
        "description": "All routes are displayed for the creationGoals app.",
        "responses": {
          "200": {
            "description": "SUCCESS, GET returned the SWAGGER API DOCUMENTATION page"
          },
          "500": {
            "description": "There was an INTERNAL SERVER ERROR while trying to GET the SWAGGER API DOCUMENTATION page"
          }
        }
      }
    },
...
    "/creationGoals/": {
      "get": {
        "summary": "GETS all the creationGoals with a status of Public ---------- (!!!OAUTH PROTECTED ROUTE!!!)",
        "description": "All public creationGoals are displayed on the creationGoals index page.",
        "responses": {
          "200": {
            "description": "SUCCESS, GET returned all Public creationGoals"
          },
          "401": {
            "description": "You are NOT AUTHORIZED to GET the Public creationGoals"
          },
          "500": {
            "description": "There was an INTERNAL SERVER ERROR while trying to GET all Public creationGoals"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "POSTS the data entered into the form page to create a new creationGoal ---------- (!!!OAUTH PROTECTED ROUTE!!!)",
        "description": "The filled out form for a new creationGoal is added to the database & the user is notified & redirected to the dashboard page.<br><b>OPTIONAL FIELDS Category Titles to Copy & Pate to JSON Body -> \"creationNumber\":, \"creationDate\":, & \"status\":</b>",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Fields to fill out.    \n(Additional OPTIONAL FIELDS you can add to the BODY BELOW = 'creationNumber':, 'creationDate':, & 'status':)",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "goal": {
                  "type": "string",
                  "example": "New added goal"
                },
                "motivator": {
                  "type": "string",
                  "example": "New added motivator"
                },
                "desire": {
                  "type": "string",
                  "example": "New added desire"
                },
                "belief": {
                  "type": "string",
                  "example": "New added belief"
                },
                "knowledge": {
                  "type": "string",
                  "example": "New added knowledge"
                },
                "plan": {
                  "type": "string",
                  "example": "New added plan"
                },
                "action": {
                  "type": "string",
                  "example": "New added action"
                },
                "victory": {
                  "type": "string",
                  "example": "New added victory"
                }
              },
              "required": "goal"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "SUCCESS, POST added a new creationGoal to the database"
          },
          "401": {
            "description": "You are NOT AUTHORIZED to POST the form page that adds a new creationGoal"
          },
          "412": {
            "description": "The PRECONDITION FAILED in the validation of the creationGoal data"
          },
          "500": {
            "description": "There was an INTERNAL SERVER ERROR while trying to POST the form page for adding a creationGoal"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}